adjusting Poll card to render data from props & UI adjustments & moving to appropriate dir
